refactor Rust sparse and add structure-only array support
