+++ tests/data/test_negative_sampling.py
@@ -112,3 +112,29 @@ def test_negative_verified_bench(rng: np.random.Generator, ml_ds: Dataset, bench
         _items = matrix.sample_negatives(users, rng=rng)
 
     benchmark(sample)
+
+
+@mark.benchmark()
+def test_negative_5000_bench(rng: np.random.Generator, ml_ds: Dataset, benchmark):
+    matrix = ml_ds.interactions().matrix()
+
+    users = rng.choice(ml_ds.user_count, 5000, replace=True)
+    users = np.require(users, "i4")
+
+    def sample():
+        _items = matrix.sample_negatives(users, rng=rng)
+
+    benchmark(sample)
+
+
+@mark.benchmark()
+def test_negative_multiple_bench(rng: np.random.Generator, ml_ds: Dataset, benchmark):
+    matrix = ml_ds.interactions().matrix()
+
+    users = rng.choice(ml_ds.user_count, 500, replace=True)
+    users = np.require(users, "i4")
+
+    def sample():
+        _items = matrix.sample_negatives(users, n=10, rng=rng)
+
+    benchmark(sample)
